// src/lib/interfaces/ListaDePecas.ts
export default interface ListaDePecas {
    camisas: string
    calca: string
    vestido: string
    terno: string
    toalha: string
    roupaDeCama: string
    blusaMoletomSueter: string
    jaquetaCasaco: string
    shorts: string
    roupasIntimas: string
    meias: string
    servicoDesejado: string
    condicoesDasPecas: string
    observacoes: string
}